Fix use of hard-coded relative paths for tempprary files.
This issue was raised on bitbucket:
https://bitbucket.org/labscript_suite/labscript_devices/pull-requests/71/

BLACS' use of hard-coded paths for tempfiles caused problems, since
they are relative paths and so are in whichever directory BLACS is running in,
which it might not have write permissions for.

This change uses full path to a random filename in the platform's proper
temporary folder instead, avoiding permissions and name collision issues.